requirements.txt
python-telegram-bot==13.15
beautifulsoup4==4.11.2
httpx==0.23.3
python-dotenv==1.0.0